Replace pkg-dir with escalade
Ref https://github.com/lukeed/escalade/

pkg-dir and many other Sindre's modules are often bloated with
dependencies. For example I have 4 instances of pkg-dir and every its
dependency in my node_modules. Would be cool to replace them with
something smaller like escalade which is gonna be used in
webpack-plugin-serve, yargs and will be used in the next major
jest release.